Move tests to src/test.
