+++ src/main/cc/wfa/virtual_people/common/field_filter/utils/BUILD.bazel
@@ -1,6 +1,9 @@
 load("@rules_cc//cc:defs.bzl", "cc_library")
 
-package(default_visibility = ["//src/main/cc/wfa/virtual_people/common/field_filter:__subpackages__"])
+package(default_visibility = [
+    "//src/main/cc/wfa/virtual_people/common/field_filter:__subpackages__",
+    "//src/test/cc/wfa/virtual_people/common/field_filter:__subpackages__"
+])
 
 _INCLUDE_PREFIX = "/src/main/cc"
 
@@ -32,21 +35,6 @@ cc_library(
     ],
 )
 
-cc_test(
-    name = "field_util_test",
-    srcs = ["field_util_test.cc"],
-    deps = [
-        ":field_util",
-        "//src/main/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
-        "@com_google_absl//absl/status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-    ],
-)
-
 cc_library(
     name = "type_convert_util",
     srcs = ["type_convert_util.cc"],
@@ -58,19 +46,6 @@ cc_library(
         "@com_google_absl//absl/status:statusor",
         "@com_google_absl//absl/strings",
         "@com_google_protobuf//:protobuf",
-    ],
-)
-
-cc_test(
-    name = "type_convert_util_test",
-    srcs = ["type_convert_util_test.cc"],
-    deps = [
-        ":type_convert_util",
-        "//src/main/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
-        "@com_google_absl//absl/status",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
     ],
 )
 
@@ -92,21 +67,6 @@ cc_library(
     ],
 )
 
-cc_test(
-    name = "integer_comparator_test",
-    srcs = ["integer_comparator_test.cc"],
-    deps = [
-        ":field_util",
-        ":integer_comparator",
-        "//src/main/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
-        "@com_google_absl//absl/status",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
-    ],
-)
-
 cc_library(
     name = "message_filter_util",
     srcs = ["message_filter_util.cc"],
@@ -119,20 +79,6 @@ cc_library(
         "@com_google_absl//absl/strings",
         "@com_google_protobuf//:protobuf",
         "@cross_media_measurement//src/main/cc/wfa/measurement/common:macros",
-    ],
-)
-
-cc_test(
-    name = "message_filter_util_test",
-    srcs = ["message_filter_util_test.cc"],
-    deps = [
-        ":message_filter_util",
-        "//src/main/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
-        "@com_google_absl//absl/status",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
     ],
 )
 
@@ -149,20 +95,5 @@ cc_library(
         "@com_google_absl//absl/strings",
         "@com_google_protobuf//:protobuf",
         "@cross_media_measurement//src/main/cc/wfa/measurement/common:macros",
-    ],
-)
-
-cc_test(
-    name = "values_parser_test",
-    srcs = ["values_parser_test.cc"],
-    deps = [
-        ":values_parser",
-        "//src/main/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
-        "@com_google_absl//absl/status",
-        "@com_google_absl//absl/status:statusor",
-        "@com_google_googletest//:gtest_main",
-        "@com_google_protobuf//:protobuf",
-        "@cross_media_measurement//src/test/cc/testutil:matchers",
-        "@cross_media_measurement//src/test/cc/testutil:status_macros",
     ],
 )
