+++ src/test/cc/wfa/virtual_people/common/field_filter/BUILD.bazel
@@ -1,0 +1,191 @@
+load("@rules_cc//cc:defs.bzl", "cc_test")
+
+package(default_visibility = ["//visibility:private"])
+
+cc_test(
+    name = "field_filter_test",
+    srcs = ["field_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "has_filter_test",
+    srcs = ["has_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "equal_filter_test",
+    srcs = ["equal_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "gt_filter_test",
+    srcs = ["gt_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "lt_filter_test",
+    srcs = ["lt_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "or_filter_test",
+    srcs = ["or_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "and_filter_test",
+    srcs = ["and_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "not_filter_test",
+    srcs = ["not_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "true_filter_test",
+    srcs = ["and_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "partial_filter_test",
+    srcs = ["partial_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "in_filter_test",
+    srcs = ["in_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "any_in_filter_test",
+    srcs = ["any_in_filter_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter",
+        "//src/main/proto/wfa/virtual_people/common:field_filter_cc_proto",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
