+++ src/test/cc/wfa/virtual_people/common/field_filter/and_filter_test.cc
@@ -12,81 +12,92 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include "absl/status/status.h"
 #include "absl/status/statusor.h"
 #include "gmock/gmock.h"
+#include "google/protobuf/text_format.h"
 #include "gtest/gtest.h"
-#include "src/main/cc/wfa/virtual_people/common/field_filter/test/test.pb.h"
 #include "src/main/proto/wfa/virtual_people/common/field_filter.pb.h"
 #include "src/test/cc/testutil/matchers.h"
 #include "src/test/cc/testutil/status_macros.h"
+#include "src/test/cc/wfa/virtual_people/common/field_filter/test/test.pb.h"
 #include "wfa/virtual_people/common/field_filter/field_filter.h"
 
 namespace wfa_virtual_people {
 namespace {
 
 using ::wfa::StatusIs;
+using ::wfa_virtual_people::test::TestProto;
 
-TEST(TrueFilterTest, TestIsMatch) {
+TEST(AndFilterTest, TestNoSubFilters) {
   FieldFilterProto field_filter_proto;
-  field_filter_proto.set_op(FieldFilterProto::TRUE);
-  ASSERT_OK_AND_ASSIGN(
-      std::unique_ptr<FieldFilter> field_filter,
-      FieldFilter::New(TestProto().GetDescriptor(), field_filter_proto));
-
-  TestProto test_proto;
-  EXPECT_TRUE(field_filter->IsMatch(test_proto));
-}
-
-TEST(TrueFilterTest, TestIsMatchNotEmptyEvent) {
-  FieldFilterProto field_filter_proto;
-  field_filter_proto.set_op(FieldFilterProto::TRUE);
-  ASSERT_OK_AND_ASSIGN(
-      std::unique_ptr<FieldFilter> field_filter,
-      FieldFilter::New(TestProto().GetDescriptor(), field_filter_proto));
-
-  TestProto test_proto;
-  test_proto.set_person_country_code("COUNTRY_1");
-  EXPECT_TRUE(field_filter->IsMatch(test_proto));
-}
-
-TEST(TrueFilterTest, TestWithName) {
-  FieldFilterProto field_filter_proto;
-  ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
-      op: TRUE
-      name: "a"
-  )pb", &field_filter_proto));
+  field_filter_proto.set_op(FieldFilterProto::AND);
   EXPECT_THAT(
       FieldFilter::New(TestProto().GetDescriptor(),
                        field_filter_proto).status(),
       StatusIs(absl::StatusCode::kInvalidArgument, ""));
 }
 
-TEST(TrueFilterTest, TestWithValue) {
+TEST(AndFilterTest, TestMatch) {
   FieldFilterProto field_filter_proto;
   ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
-      op: TRUE
-      value: "1"
-  )pb", &field_filter_proto));
-  EXPECT_THAT(
-      FieldFilter::New(TestProto().GetDescriptor(),
-                       field_filter_proto).status(),
-      StatusIs(absl::StatusCode::kInvalidArgument, ""));
-}
-
-TEST(TrueFilterTest, TestWithSubFilters) {
-  FieldFilterProto field_filter_proto;
-  ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
-      op: TRUE
+      op: AND
       sub_filters {
+        name: "a.b.int32_value"
         op: EQUAL
-        name: "a"
+        value: "1"
+      }
+      sub_filters {
+        name: "a.b.int64_value"
+        op: EQUAL
         value: "1"
       }
   )pb", &field_filter_proto));
-  EXPECT_THAT(
-      FieldFilter::New(TestProto().GetDescriptor(),
-                       field_filter_proto).status(),
-      StatusIs(absl::StatusCode::kInvalidArgument, ""));
+  ASSERT_OK_AND_ASSIGN(
+      std::unique_ptr<FieldFilter> field_filter,
+      FieldFilter::New(TestProto().GetDescriptor(), field_filter_proto));
+
+  TestProto test_proto;
+  ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
+      a {
+        b {
+          int32_value: 1
+          int64_value: 1
+        }
+      }
+  )pb", &test_proto));
+  EXPECT_TRUE(field_filter->IsMatch(test_proto));
+}
+
+TEST(AndFilterTest, TestNotMatch) {
+  FieldFilterProto field_filter_proto;
+  ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
+      op: AND
+      sub_filters {
+        name: "a.b.int32_value"
+        op: EQUAL
+        value: "1"
+      }
+      sub_filters {
+        name: "a.b.int64_value"
+        op: EQUAL
+        value: "1"
+      }
+  )pb", &field_filter_proto));
+  ASSERT_OK_AND_ASSIGN(
+      std::unique_ptr<FieldFilter> field_filter,
+      FieldFilter::New(TestProto().GetDescriptor(), field_filter_proto));
+
+  TestProto test_proto;
+  ASSERT_TRUE(google::protobuf::TextFormat::ParseFromString(R"pb(
+      a {
+        b {
+          int32_value: 1
+          int64_value: 2
+        }
+      }
+  )pb", &test_proto));
+  EXPECT_FALSE(field_filter->IsMatch(test_proto));
 }
 
 }  // namespace
