+++ src/test/cc/wfa/virtual_people/common/field_filter/utils/BUILD.bazel
@@ -1,0 +1,75 @@
+load("@rules_cc//cc:defs.bzl", "cc_test")
+
+package(default_visibility = ["//visibility:private"])
+
+cc_test(
+    name = "field_util_test",
+    srcs = ["field_util_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:field_util",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "type_convert_util_test",
+    srcs = ["type_convert_util_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:type_convert_util",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+    ],
+)
+
+cc_test(
+    name = "integer_comparator_test",
+    srcs = ["integer_comparator_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:field_util",
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:integer_comparator",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "message_filter_util_test",
+    srcs = ["message_filter_util_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:message_filter_util",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
+
+cc_test(
+    name = "values_parser_test",
+    srcs = ["values_parser_test.cc"],
+    deps = [
+        "//src/main/cc/wfa/virtual_people/common/field_filter/utils:values_parser",
+        "//src/test/cc/wfa/virtual_people/common/field_filter/test:test_cc_proto",
+        "@com_google_absl//absl/status",
+        "@com_google_absl//absl/status:statusor",
+        "@com_google_googletest//:gtest_main",
+        "@com_google_protobuf//:protobuf",
+        "@cross_media_measurement//src/test/cc/testutil:matchers",
+        "@cross_media_measurement//src/test/cc/testutil:status_macros",
+    ],
+)
