+++ src/test/cc/wfa/virtual_people/common/field_filter/utils/values_parser_test.cc
@@ -20,9 +20,9 @@
 #include "google/protobuf/descriptor.h"
 #include "google/protobuf/message.h"
 #include "gtest/gtest.h"
-#include "src/main/cc/wfa/virtual_people/common/field_filter/test/test.pb.h"
 #include "src/test/cc/testutil/matchers.h"
 #include "src/test/cc/testutil/status_macros.h"
+#include "src/test/cc/wfa/virtual_people/common/field_filter/test/test.pb.h"
 
 namespace wfa_virtual_people {
 namespace {
